+++ front-page.php
@@ -31,17 +31,17 @@
 							<p>From landlords and tenants to buyers and sellers, we provide consultative commercial real estate solutions focused on building <strong>long-term relationships</strong> with you and <strong>long-term success</strong> for your business.</p>
 						</div>
 						<div class="selling-point m-all t-1of3 d-1of3">
-							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icons-book-knowledge.svg" alt="Real Estate Knowledge">
+							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-book-knowledge.svg" alt="Real Estate Knowledge">
 							<h5>A wealth of knowledge.</h5>
 							<p>With over 600 years of combined experience, our team of experts provides you with the latest data, marketing strategies and support you deserve.</p>
 						</div>
 						<div class="selling-point m-all t-1of3 d-1of3">
-							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icons-us-map-love-indy.svg" alt="Real Estate Knowledge">
+							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-us-map-love-indy.svg" alt="Real Estate Knowledge">
 							<h5>Locally-owned.</h5>
 							<p>As the largest, locally-owned firm in Indianapolis, we are deeply-connected experts who care about the success of our city. We are proud of our roots!</p>
 						</div>
 						<div class="selling-point m-all t-1of3 d-1of3 right">
-							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icons-focused-on-you.svg" alt="Real Estate Knowledge">
+							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-focused-on-you.svg" alt="Real Estate Knowledge">
 							<h5>Free to focus on you.</h5>
 							<p>We don&#39;t answer to stockholders. We aren&#39;t an oversized corporation that only cares about boosting our numbers. We care about your numbers.</p>
 						</div>
@@ -57,7 +57,7 @@
 				<section id="need-help">
 					<div class="wrap">
 						<h2>Can&#39;t find what you&#39;re looking for?</h2>
-						<p>We want to ensure we make your life as easy as possible. </br>The best way we can serve you is <a id="cta-underline-white" href="https://www.getfeedback.com/r/TXPTw5oV">sending us a message</a> or <a id="cta-underline-white" href="tel:317-663-6000" role="telephone">giving us a call.</a> No hoops.</p>
+						<p>We want to ensure we make your life as easy as possible. </br>The best way we can serve you is <a class="cta-underline-white" href="https://www.getfeedback.com/r/TXPTw5oV">sending us a message</a> or <a class="cta-underline-white" href="tel:317-663-6000" role="telephone">giving us a call.</a> No hoops.</p>
 						<p>We <i class="fa fa-heart" aria-hidden="true"></i> feedback! Let us know how we can make our site better!</p>
 					</div>
 				</section>