+++ header.php
@@ -161,53 +161,19 @@
 				get_template_part( 'library/partials/header-home' );
 			
 			} elseif( get_post_type() == "employee") {
-				// Header for individual blog posts, case studies, and research reports
 				get_template_part('library/partials/header-none');
+
+			} elseif( get_post_type() == "listing") {
+				get_template_part('library/partials/header-listing');
 
 			} elseif( is_404() ) {
 				get_template_part('library/partials/header-404');
 
-			} elseif ( is_single() ) { ?>
-			<!-- Header for Single.php and works on listing pages-->
-			<div id="single-header" style="background-image: url('<?php echo $bannerimg; ?>'); background-repeat: no-repeat; background-size: cover;" role="banner" itemscope itemtype="http://schema.org/WPHeader">
+			} elseif ( is_single() ) {
+				get_template_part('library/partials/header-single');
 
-					<div class="m-padding cf">
-						
-						<div class="m-all pull-r-1of12 pull-l-1of12 t-1of2 d-11of12 cf">
-							
-							<div class="tag blue">
-							<?php echo get_the_term_list( $post->ID, 'specialty', '', '', ''); ?>
-							</div>
-
-							<?php if ( is_tax('service') ) { ?>
-								<div class="tag gray">
-							<?php echo get_the_term_list( $post->ID, 'service', '', '', ''); ?>
-							</div>
-							<?php } ?>
-							
-							<h1 class="cs-h1"><?php the_title(); ?></h1>
-							<p class="subtitle">
-							<?php echo types_render_field( "tagline", array( 'raw' => true)); ?>
-							</p>
-
-						</div>
-
-						<?php if ( types_render_field( "report-pdf" ) != null ) { ?>
-							<a title="Download Print-friendly Version" class="download-icon pull-l-1of12" href="<?php echo types_render_field( "report-pdf", array( $post->ID, 'raw' => true ) ); ?>" alt="Download the PDF" target="_blank"></a>
-						<?php } 
-
-						if ( types_render_field( "downloadable-pdf" ) != null ) { ?>
-							<a title="Download the PDF" class="download-icon pull-l-1of12" href="<?php echo types_render_field( "downloadable-pdf", array( $post->ID, 'raw' => true)); ?>" alt="Download the PDF" target="_blank"></a>
-						<?php } ?>
-					</div>
-				</div>
-
-			<?php } elseif ( is_single() ) {
-				// Get header based on post type
-	 			get_template_part('library/partials/header', get_post_type() );
-			
 			} else { ?>
-			<!-- Header for all other pages -->
+			<!-- Header 6 - Header for all other pages -->
 				<div id="page-header" style="background-image: url('<?php echo $bannerimg; ?>');" role="banner" itemscope itemtype="http://schema.org/WPHeader">
 					<div class="callout cf">
 						<div class="m-all t-1of2 d-1of2 center cf">
@@ -215,6 +181,11 @@
 								<?php the_title(); ?>
 							</h1>
 							<p class="subtitle"><?php echo types_render_field( "tagline", array( 'raw' => true)); ?></p>
+							
+							<?php if ( is_home() ) { ?>
+								<a title="Read More" class="cta-border-green btn-minify" href="<?php echo esc_url( get_permalink( ) ); ?>">Read More</a>
+							<?php } ?>
+							
 						</div>
 					</div>
 				</div>