+++ single-listing.php
@@ -16,6 +16,26 @@
 
 			<section class="topic cf">
 				<div class="pull-l-1of12 pull-r-1of12 pad-top cf">
+					
+					<div class="m-all t-1of2 d-1of2 cf">
+						<div class="center">
+							<iframe
+						  width="100%"
+						  height="350"
+						  frameborder="0"
+						  src="https://www.google.com/maps?key=AIzaSyAVyvyZSwaa7p6T0z0eIYDLeSb2YNauVs8
+						    &q=<?php echo(types_render_field( "address", array( 'raw' => true) )); ?>&maptype=satellite&output=embed" allowfullscreen>
+							</iframe>
+						</div>
+						
+
+						<a href="<?php echo(types_render_field( "property-flyer", array( 'raw' => true) )); ?>" class="cta-border-green btn-minify" target="_blank">Download Property Flyer</a>
+						
+
+						<a href="<?php echo ( types_render_field( "google-maps-url", array( 'raw' => true) )); ?>" class="cta-border-green btn-minify" target="_blank">View on Google Maps</a>
+						
+					</div>
+
 					<div class="m-all t-1of2 d-1of2 cf">
 						
 						<?php if ( types_render_field( "flyer-headline" ) != null ) { ?>
@@ -33,14 +53,7 @@
 						<?php } ?>
 					</div>
 				
-					<div class="m-all t-1of2 d-1of2 pad-top cf">
-
-						<p class="m-1of2 t-1of2 d-1of2"><a href="<?php echo(types_render_field( "property-flyer", array( 'raw' => true) )); ?>" class="cta-border-green" target="_blank">Download Property Flyer</a>
-						</p>
-
-						<p class="m-1of2 t-1of2 d-1of2"><a href="<?php echo ( types_render_field( "google-maps-url", array( 'raw' => true) )); ?>" class="cta-border-green" target="_blank">View on Google Maps</a>
-						</p>
-					</div>
+					
 				</div>
 			</section>